+++ src/Core/planetClass.java
@@ -37,6 +37,7 @@ public class planetClass extends planetCore implements Serializable {
     private colonyCore planetColony;
     private double resources;
     private int colonyID;
+    private boolean homePlanet;
 
     //generates a random planet
     public planetClass(starClass parentStar, int planetNumber) {
@@ -46,9 +47,9 @@ public class planetClass extends planetCore implements Serializable {
         this.planetRadius = calculateSize(parentStar.getStarRadius(), this.distanceFromStar); //determines the radius of the planet in km
         this.isTidallyLocked = checkTidalLock(this.planetRadius, parentStar.getStarRadius()); //runs a check to determine tidal lock
         //this.isInHabitableZone = isInHabitableZone(this.distanceFromStar, parentStar.getHabitableZoneMax(), parentStar.getHabitableZoneMin());
-        this.isPlanetHabitable = determineHabitability(arrayLoc); //whether or not the planet can be colonized
-        this.planetType = determinePlanetClass(this.distanceFromStar, this.planetRadius, this.isTidallyLocked, this.isPlanetHabitable, this.parentStar); //determines the planet's class
+        this.planetType = determinePlanetClass(this.distanceFromStar, this.planetRadius, this.isTidallyLocked, this.parentStar.isStarHabitable(), this.parentStar); //determines the planet's class
         this.arrayLoc = getPlanetFromID(planetType); //Gets the planet's array location from the ID.
+        this.isPlanetHabitable = determineHabitability(arrayLoc); //whether or not the planet can be colonized
         this.isHabited = false; //by default, the planet is not currently colonized
         this.planetName = this.parentStar.getStarName() + " " + this.planetNumber;
         //System.out.println("Planet '" + listOfPlanets.get(arrayLoc).getClassName() + "' (ID" + planetType + ") successfully generated.");
@@ -56,7 +57,7 @@ public class planetClass extends planetCore implements Serializable {
     }
 
     //generates a pre-defined planet
-    public planetClass(starClass parentStar, String planetName, int planetType, int planetNumber, long distanceFromStar, int planetRadius, boolean isInHabitableZone, boolean isHabited, boolean isPlanetHabitable, boolean isTidallyLocked, int numOfMoons) {
+    public planetClass(starClass parentStar, String planetName, int planetType, int planetNumber, long distanceFromStar, int planetRadius, boolean isInHabitableZone, boolean isHabited, boolean isPlanetHabitable, boolean isTidallyLocked, int numOfMoons, boolean homePlanet) {
         this.parentStar = parentStar;
         this.planetName = planetName;
         this.planetType = planetType;
@@ -69,6 +70,7 @@ public class planetClass extends planetCore implements Serializable {
         this.isPlanetHabitable = isPlanetHabitable;
         this.isTidallyLocked = isTidallyLocked;
         this.numOfMoons = numOfMoons;
+        this.homePlanet = homePlanet;
         //System.out.println("Predefined planet '" + listOfPlanets.get(arrayLoc).getClassName() + " [" + this.planetName + "] (ID" + planetType + ") successfully generated in the " + parentStar.getStarName() + " system.");
 
     }
@@ -90,6 +92,7 @@ public class planetClass extends planetCore implements Serializable {
     public int getArrayLoc() { return this.arrayLoc; }
     public String getPlanetClassName() { return planetClass.listOfPlanets.get(arrayLoc).getClassName(); }
     public String getPlanetClassDesc() { return planetClass.listOfPlanets.get(arrayLoc).getClassDesc(); }
+    public boolean isHomePlanet() { return this.homePlanet; }
 
     /** Setter methods **/